put cleanup methods together
